pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "setuptools-scm>=8.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tritonparse"
dynamic = ["version"]
dependencies = []
requires-python = ">=3.10"
description = "TritonParse: A Compiler Tracer, Visualizer, and mini-Reproducer Generator for Triton Kernels"
readme = "README.md"
authors = [
    { name="Yueming Hao", email="yhao@meta.com" },
]
license = "BSD-3-Clause"
license-files = ["LICENSE*"]


[project.optional-dependencies]
triton = [
    "triton>3.3.1",
]
pytorch-triton = [
    "pytorch-triton>=3.4.0",
]
test = [
    "coverage>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["tritonparse*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ufmt]
formatter = "black"
sorter = "usort"

[tool.usort]
first_party_detection = false

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[project.urls]
"Homepage" = "https://github.com/meta-pytorch/tritonparse"
